Modularized the application
