+++ pdf2tree/__init__.py
@@ -1,0 +1,1 @@
+from .pipeline import build_tree_direct as pdf_to_tree
